[BPK-4276] Restructure iOS pages
