+++ docs/src/pages/PanelPage/IOSPanelPage.js
