+++ docs/src/pages/PanelPage/PanelPage.js
@@ -23,9 +23,9 @@ import React from 'react';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import Android from '../AndroidPanelPage';
-import IOS from '../IOSPanelPage';
 import Web from '../WebPanelsPage';
 
+import IOS from './IOSPanelPage';
 import Native from './NativePanelsPage';
 
 const Page = () => (
