+++ docs/src/pages/RatingPage/IOSRatingPage.js
