+++ docs/src/pages/RatingPage/RatingPage.js
@@ -23,9 +23,9 @@ import React from 'react';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import AndroidRating from '../AndroidRatingPage';
-import IOSRating from '../IOSRatingPage';
 import WebRating from '../WebRatingPage';
 
+import IOSRating from './IOSRatingPage';
 import NativeRating from './NativeRatingPage';
 
 const Page = () => (
