+++ docs/src/pages/SpinnerPage/SpinnerPage.js
@@ -23,9 +23,9 @@ import React from 'react';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import Android from '../AndroidSpinnerPage';
-import IOS from '../IOSSpinnerPage';
 import Web from '../WebSpinnersPage';
 
+import IOS from './IOSSpinnerPage';
 import Native from './NativeSpinnerPage';
 
 const blurb = [
