+++ docs/src/pages/StarRatingPage/IOSStarRatingPage.js
