+++ docs/src/pages/SwitchPage/IOSSwitchPage.js
