+++ docs/src/pages/SwitchPage/SwitchPage.js
@@ -21,9 +21,9 @@ import React from 'react';
 import IntroBlurb from '../../components/IntroBlurb';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import AndroidSwitchPage from '../AndroidSwitchPage';
-import IOSSwitchPage from '../IOSSwitchPage';
 import WebSwitchPage from '../WebSwitchPage';
 
+import IOSSwitchPage from './IOSSwitchPage';
 import NativeSwitchPage from './NativeSwitchPage';
 
 const blurb = [
