+++ docs/src/pages/ToastPage/ToastPage.js
@@ -23,7 +23,8 @@ import React from 'react';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import AndroidPage from '../AndroidToastPage';
-import IOSPage from '../IOSToastPage';
+
+import IOSPage from './IOSToastPage';
 
 const Page = () => (
   <DocsPageWrapper
